TestHelper classes for shared types
