+++ test/Microsoft.AspNetCore.Server.KestrelTests/TestHelpers/PassThroughConnectionFilter.cs
@@ -1,0 +1,19 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Server.Kestrel.Filter;
+using Microsoft.AspNetCore.Server.Kestrel.Infrastructure;
+
+namespace Microsoft.AspNetCore.Server.KestrelTests
+{
+
+    public class PassThroughConnectionFilter : IConnectionFilter
+    {
+        public Task OnConnectionAsync(ConnectionFilterContext context)
+        {
+            context.Connection = new LoggingStream(context.Connection, new TestApplicationErrorLogger());
+            return TaskUtilities.CompletedTask;
+        }
+    }
+}
